backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel",email = "daniellimafreita@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djoser[crypto] (>=2.3.1,<3.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
